requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.3.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
pyotp==2.9.0
qrcode==7.4.2

# Background Tasks
# celery[redis]==5.3.4
# redis==5.0.1

# Email
boto3==1.34.34

# File Upload
aiofiles==23.2.1
Pillow==10.2.0

# HTTP Client
httpx==0.26.0

# Monitoring & Logging
sentry-sdk[fastapi]==1.40.0
python-json-logger==2.0.7

# Security & Rate Limiting
slowapi==0.1.9
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
